requirements.txt
openai>=1.0.0
langchain>=1.0.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
chromadb>=0.4.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pypdf>=3.0.0
tiktoken>=0.5.0
streamlit>=1.28.0